Don't restrict shares to 3 character extensions
Whenever sharing, the set of files used for the upload was calculated by
removing the last 3 characters of the path and appending the new file
extensions. This meant that every level must have an associated file
extension of only 3 characters. Switch to stripping off the whole of the
original extension and appending the new one instead.