Wardrobe: Move wardrobe screen app from Dashboard
The dashboard onboarding requires the profile to open on the wardrobe
selection screen with the user in edit mode. To do this a quick solution
of adding a new executable was enacted, resulting in not only a new
executable but also it being committed into the dashboard repo. Resolve
the latter of these issues with the hope that the former will be
resolved sooner rather than later.