World Hooks: Isolate apps launched from Kano World
When an app is launched via the Kano XDG scheme from the browser, the
browser is the parent of the process and hence all child processes. This
poses a problem in Sonic Pi when the share window is launched and causes
the window to dump its core, however this is only a problem if the
browser is the parent of the process. As such, migrate all browser-based
launches to be run through SystemD.