Tracking: Prevent failure when resuming nothing
If, when the sessions are paused, there are no sessions to pause then an
empty file is created to store this. Upon resuming, this file is parsed
as a JSON and it is determined to be the empty document {}. This causes
the resuming to trace back. Prevent this from happening by catching and
ignoring this exception.