Tracking: Prevent errors when double resuming sessions
When resuming sessions, an attempt to remove the file which holds the
sessions is made. If the sessions are resumed twice in succession then
two attempts are made but by the second attempt the file has already
been removed. Prevent errors by checking for the file's existence before
trying to remove it.