CI: Run tests and generate coverage report on build
Add a Jenkins build step to run the test suite and generate the coverage
report. Facilitate this change by migrating all non-standard imports in
the library to within the test functions themselves so that one failing
import doesn't prevent the whole suite from running. Also tag tests
which require Gtk so that they can be skipped when running on the
server.

Output a JUnit file with the test results and use this to communicate
the test results from the PyTest and Behave suites.